requirements.txt
streamlit
python-dotenv
google-generativeai
huggingface_hub
gtts
sentence-transformers
scikit-learn
langdetect
deep-translator